re_viewport_blueprint
Module ui

UI utilities related to the viewport blueprint.

Current this is mainly the add view or container modal.

Modules

add_view_or_container_modal 🔒

Modal for adding a new view of container to an existing target container.

Statics

ADD_VIEW_OR_CONTAINER_MODAL 🔒

Functions

add_view_or_container_modal_ui

show_add_view_or_container_modal