Crate re_viewport_blueprint

Rerun Viewport Blueprint

This crate provides blueprint (i.e. description) for how to render the viewport.

Modules

container 🔒

space_view 🔒

space_view_contents 🔒

tree_actions 🔒

ui

UI utilities related to the viewport blueprint.

view_properties 🔒

viewport_blueprint 🔒

Structs

ContainerBlueprint

The native version of a re_types_blueprint::blueprint::archetypes::ContainerBlueprint.

SpaceViewBlueprint

A view of a space.

SpaceViewContents

Data to be added to a space view, built from a blueprint_archetypes::SpaceViewContents.

ViewProperty

Utility for querying view properties.

ViewportBlueprint

Describes the layout and contents of the Viewport Panel.

Enums

TreeAction

Mutation actions to perform on the tree at the end of the frame. These messages are sent by the mutation APIs from crate::ViewportBlueprint.

ViewPropertyQueryError

Constants

VIEWPORT_PATH

Functions

container_kind_from_egui

Converts a [egui_tiles::ContainerKind] into a re_types_blueprint::blueprint::components::ContainerKind.

container_kind_to_egui

Converts a re_types_blueprint::blueprint::components::ContainerKind into a [egui_tiles::ContainerKind].

default_created_space_views

List out all space views we generate by default for the available data.

entity_path_for_view_property

Entity path in the blueprint store where all components of the given view property archetype are stored.