Crate re_viewport_blueprint

Rerun Viewport Blueprint

This crate provides blueprint (i.e. description) for how to render the viewport.

Modules

auto_layout 🔒

Code for automatic layout of views.

container 🔒

entity_add_info 🔒

Utilities for determining if an entity can be added to a view.

test_context_ext

ui

UI utilities related to the viewport blueprint.

view 🔒

view_contents 🔒

view_properties 🔒

viewport_blueprint 🔒

viewport_command 🔒

Structs

ContainerBlueprint

The native version of a re_types::blueprint::archetypes::ContainerBlueprint.

DataQueryPropertyResolver

EntityAddInfo

ViewBlueprint

A view of a space.

ViewContents

Data to be added to a view, built from a blueprint_archetypes::ViewContents.

ViewProperty

Utility for querying view properties.

ViewportBlueprint

Describes the layout and contents of the Viewport Panel.

Enums

CanAddToView

Describes if an entity path can be added to a view.

ViewPropertyQueryError

ViewportCommand

Mutation actions to perform on the viewport tree at the end of the frame.

Constants

VIEWPORT_PATH

The entity path of the viewport blueprint in the blueprint store.

Functions

container_kind_from_egui

Converts a [egui_tiles::ContainerKind] into a re_types::blueprint::components::ContainerKind.

container_kind_to_egui

Converts a re_types::blueprint::components::ContainerKind into a [egui_tiles::ContainerKind].

create_entity_add_info

default_created_views

List out all views we generate by default for the available data.

entity_path_for_view_property

Entity path in the blueprint store where all components of the given view property archetype are stored.

tree_simplification_options